bin/stops.sql: provide stop numbers for stops table.
This script replaces the stops table with a version having stop
numbers and sorted by trip, stop & type. It has all the indices
and foreign keys on it that the old table does.

Don't bother making indices on the original stops table since it's
just going to get deleted anyway.